requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.9.0=py36_0
astroid=2.4.2=py36_0
blas=1.0=mkl
blinker=1.4=py_1
brewer2mpl=1.4.1=py_3
brotlipy=0.7.0=py36h8c4c3a4_1000
c-ares=1.15.0=h516909a_1001
ca-certificates=2020.6.24=0
cachetools=4.1.1=py_0
certifi=2020.6.20=py36_0
cffi=1.11.5=py36_0
chardet=3.0.4=py36h9f0ad1d_1006
click=7.1.2=pyh9f0ad1d_0
cryptography=2.9.2=py36h45558ae_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=py36_0
dbus=1.13.16=hb2f20db_0
expat=2.2.9=he6710b0_2
fontconfig=2.13.0=h9420a91_0
freetype=2.10.2=h5ab3b9f_0
future=0.18.2=pypi_0
glib=2.65.0=h3eb4bd4_0
google-auth=1.17.2=py_0
google-auth-oauthlib=0.4.1=py_2
grpcio=1.30.0=py36h769ab6c_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb31296c_0
icu=58.2=he6710b0_3
idna=2.10=pyh9f0ad1d_0
importlib-metadata=1.7.0=py36h9f0ad1d_0
intel-openmp=2020.1=217
isort=4.3.21=py36_0
joblib=0.16.0=py_0
jpeg=9b=h024ee3a_2
kiwisolver=1.2.0=py36hfd86e86_0
lazy-object-proxy=1.4.3=py36h7b6447c_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.12.3=h8b12597_1
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=he19cac6_1
lz4-c=1.9.2=he6710b0_0
markdown=3.2.2=py_0
matplotlib=3.2.2=0
matplotlib-base=3.2.2=py36hef1b27d_0
mccabe=0.6.1=py36_1
mkl=2020.1=217
mkl-service=2.3.0=py36he904b0f_0
mkl_fft=1.1.0=py36h23d657b_0
mkl_random=1.1.1=py36h0573a6f_0
ncurses=6.2=he6710b0_1
ninja=1.9.0=py36hfd86e86_0
numpy=1.18.5=py36ha1c710e_0
numpy-base=1.18.5=py36hde5b4d6_0
oauthlib=3.0.1=py_0
olefile=0.46=py36_0
openssl=1.1.1g=h7b6447c_0
pcre=8.44=he6710b0_0
pillow=7.2.0=py36hb39fc2d_0
pip=20.1.1=py36_1
protobuf=3.12.3=py36h831f99a_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=pyh9f0ad1d_2
pyjwt=1.7.1=py_0
pylint=2.5.3=py36_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=py_0
pyqt=5.9.2=py36h05f1152_2
pysocks=1.7.1=py36h9f0ad1d_1
python=3.6.10=h7579374_2
python-dateutil=2.8.1=py_0
python_abi=3.6=1_cp36m
pytorch=1.5.1=py3.6_cuda10.2.89_cudnn7.6.5_0
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
requests=2.24.0=pyh9f0ad1d_0
requests-oauthlib=1.2.0=py_0
rsa=4.6=pyh9f0ad1d_0
scikit-learn=0.23.1=py36h423224d_0
scipy=1.5.0=py36h0b6359f_0
setuptools=45.2.0=py36_0
sip=4.19.8=py36hf484d3e_0
six=1.15.0=py_0
sqlite=3.32.3=h62c20be_0
tensorboard=2.2.2=py_0
tensorboard-plugin-wit=1.6.0=pyh9f0ad1d_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hbc83047_0
toml=0.10.1=py_0
torchprof=1.1.1=dev_0
torchvision=0.6.1=py36_cu102
tornado=6.0.4=py36h7b6447c_1
tqdm=4.47.0=py_0
typed-ast=1.4.1=py36h7b6447c_0
urllib3=1.25.9=py_0
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.34.2=py36_0
wrapt=1.11.2=py36h7b6447c_0
xz=5.2.5=h7b6447c_0
zipp=3.1.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.4=h0b5b093_3
